+++ Sources/SuperwallKit/Paywall/Presentation/PaywallInfo.swift
@@ -3,8 +3,6 @@
 //  Superwall
 //
 //  Created by Yusuf Tör on 28/02/2022.
-//
-// swiftlint:disable file_length
 
 import Foundation
 import StoreKit